pyproject.toml
[project]
name = "control-cli"
version = "0.1.0"
description = "CLI de CONTROL: Gestión agnóstica de Agentes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "pyyaml>=6.0.1",
    "rich>=13.0.0",
]

[project.scripts]
control = "control.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"